requirements.txt
appdirs==1.4.4
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
curl_cffi==0.12.0
dateutils==0.6.12
et_xmlfile==2.0.0
frozendict==2.4.6
gevent==25.5.1
greenlet==3.2.3
html5lib==1.1
idna==3.10
logzero==1.7.0
lxml==6.0.0
more-itertools==10.7.0
multitasking==0.0.12
nsepy==0.8
nsetools==2.0.1
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
premailer==3.10.0
protobuf==6.31.1
pycparser==2.22
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
smartapi-python==1.5.5
soupsieve==2.7
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
websocket==0.2.1
websockets==15.0.1
yagmail==0.15.293
yfinance==0.1.96
zope.event==5.1.1
zope.interface==7.2
